[refactor] 상품 카드 컴포넌트 분리
