+++ libs/components-web-kkshow/src/lib/search/SearchResultGoodsSection.tsx
@@ -1,58 +1,12 @@
-import { Box, LinkBox, LinkOverlay, Stack, Text, useBoolean } from '@chakra-ui/react';
+import { Box, useBreakpointValue } from '@chakra-ui/react';
 import { SearchResultItem } from '@project-lc/shared-types';
-import { motion } from 'framer-motion';
 import { Pagination, Scrollbar } from 'swiper';
 import { Swiper, SwiperSlide } from 'swiper/react';
-import NextLink from 'next/link';
-import { ArrowForwardIcon, Icon } from '@chakra-ui/icons';
+import { GoodsCard } from './GoodsCard';
 import {
   SearchResultEmptyText,
   SearchResultSectionContainer,
 } from './SearchResultSectionContainer';
-
-const variants = {
-  scale: { scale: 1.1 },
-  normal: { scale: 1.0 },
-};
-
-/** 검색결과 - 상품아이템 => 쇼핑탭에서 작성한 상품컴포넌트와 비슷할듯하다 => 확인 후 수정필요
- */
-function GoodsCard({ item }: { item: SearchResultItem }): JSX.Element {
-  const [mouseOver, setMouseOver] = useBoolean(false);
-  return (
-    <LinkBox onMouseOver={setMouseOver.on} onMouseOut={setMouseOver.off}>
-      <NextLink href={item.linkUrl} passHref>
-        <LinkOverlay isExternal={item.linkUrl.includes('http')}>
-          <Stack minW="132px">
-            <Box rounded="xl" overflow="hidden" position="relative">
-              <motion.img
-                src={item.imageUrl}
-                initial="normal"
-                animate={mouseOver ? 'scale' : 'normal'}
-                variants={variants}
-                style={{ objectFit: 'cover' }}
-              />
-              <Icon
-                as={ArrowForwardIcon}
-                position="absolute"
-                right={2}
-                bottom={2}
-                color="white"
-                rounded="full"
-                bg="rgba(0,0,0,0.4)"
-                p={1}
-                boxSize="2em"
-              />
-            </Box>
-            <Text fontFamily="Gmarket Sans" textAlign="center">
-              {item.title}
-            </Text>
-          </Stack>
-        </LinkOverlay>
-      </NextLink>
-    </LinkBox>
-  );
-}
 
 export interface SearchResultGoodsSectionProps {
   data: SearchResultItem[];
@@ -60,11 +14,14 @@ export interface SearchResultGoodsSectionProps {
 export function SearchResultGoodsSection({
   data,
 }: SearchResultGoodsSectionProps): JSX.Element {
+  // 모바일에서는 4개, 데스트탑에서는 최대 6개 표시 -> 나머지는 더보기 눌러서 별도 페이지에서 확인하도록
+
+  const displayLimitOnSearchResultPage = useBreakpointValue({ base: 4, md: 6 });
   return (
     <SearchResultSectionContainer title="상품" resultCount={data.length}>
       {data.length > 0 ? (
         <>
-          {/* (breakpoint: md 이상 기준) 상품 검색결과 목록 */}
+          {/* (breakpoint: md 이상 기준) */}
           <Box display={{ base: 'none', md: 'block' }}>
             <Swiper
               scrollbar
@@ -72,7 +29,7 @@ export function SearchResultGoodsSection({
               slidesPerView="auto"
               spaceBetween={30}
             >
-              {data.map((item, index) => {
+              {data.slice(0, displayLimitOnSearchResultPage).map((item, index) => {
                 const key = `${item.title}_${index}`;
                 return (
                   <SwiperSlide key={key} style={{ width: '20%', paddingBottom: '32px' }}>
@@ -82,7 +39,7 @@ export function SearchResultGoodsSection({
               })}
             </Swiper>
           </Box>
-          {/* (breakpoint: md 미만 기준) 상품 검색결과 목록 - 그리드 */}
+          {/* (breakpoint: md 미만 기준) - 그리드 */}
           <Box display={{ base: 'block', md: 'none' }} />
         </>
       ) : (
